==> scripts/01_downloading_files.ipynb <==
Aidan Coyle

afcoyle@uw.edu

QERM 514, Spring 2021

Written 2021/05/12, last edited 2021/05/12

Final Project

## Script 1: Download directory containing all survey data files

# Download all data from folder contained on Gannet
!wget --no-check-certificate --no-parent --recursive --reject "index.html" https://gannet.fish.washington.edu/hematodinium/Hemato%20samples/BCS%20Extraction%20Plate%20maps%202005_2013/ADFG%20Tanner%20Survey%20Data/

# Copy downloaded files into data directory
!mv gannet.fish.washington.edu/hematodinium/Hemato\ samples/BCS\ Extraction\ Plate\ maps\ 2005_2013/ADFG\ Tanner\ Survey\ Data/* ../data

# Remove remnant folder in our scripts directory
!rm -r gannet.fish.washington.edu/

# Change into data directory
%cd ../data

# Run a shell script that changes all spaces to underscores in filenames
! ../scripts/remove_spaces.sh

# Move back into scripts directory
%cd ../scripts

# Remove files that don't contain relevant data
!rm ../data/index.html*
!rm ../data/2012_ADFG_Summer_Fall_data_RAW.xlsx
!rm ../data/2011_Tanner_Leg_1_Skipper_Data_RAW.xlsx
!rm ../data/2010_ADFG_SEAK_Fall_Tanner_Survey_Data_RAW2.xls
!rm ../data/2008_ADFG_SEAK_Summer_Survey_bairdi_RAW2.xls

# List all remaining files
!ls ../data/
